// src/types/car-left-right.ts
/**
 * CarLeftRight proximity indicator enum.
 *
 * Maps to iRacing's irsdk_CarLeftRight values.
 * @see https://sajax.github.io/irsdkdocs/telemetry/carleftright/
 */
export enum CarLeftRight {
  Off = 0,
  Clear = 1,
  CarLeft = 2,
  CarRight = 3,
  CarLeftRight = 4,
  Cars2Left = 5,
  Cars2Right = 6,
}